+++ src/hooks.cpp
@@ -27,4 +27,10 @@ void script::initAllHooks() {
     CREATE_HOOK_WITH_CHECK_FOR(this->mainClass, "MenuLayer::updateUserProfileButton", "MenuLayer_updateUserProfileButton", wrapper::MenuLayer_updateUserProfileButton, geode::base::get() + 0x320330, Default, MenuLayer_updateUserProfileButtonHook)
     CREATE_HOOK_WITH_CHECK_FOR(this->mainClass, "MenuLayer::willClose", "MenuLayer_willClose", wrapper::MenuLayer_willClose, geode::base::get() + 0x3211f0, Default, MenuLayer_willCloseHook)
     CREATE_HOOK_WITH_CHECK_FOR(this->mainClass, "MenuLayer::init", "MenuLayer_init", wrapper::MenuLayer_init, geode::base::get() + 0x31ebd0, Default, MenuLayer_initHook)
+
+    CREATE_HOOK_WITH_CHECK_FOR(this->mainClass, "LoadingLayer::scene", "LoadingLayer_scene", wrapper::LoadingLayer_scene, geode::base::get() + 0x3194f0, Default, LoadingLayer_sceneHook)
+    // CREATE_HOOK_WITH_CHECK_FOR(this->mainClass, "LoadingLayer::getLoadingString", "LoadingLayer_getLoadingString", wrapper::LoadingLayer_getLoadingString, geode::base::get() + 0x31a930, Default, LoadingLayer_getLoadingStringHook)
+    CREATE_HOOK_WITH_CHECK_FOR(this->mainClass, "LoadingLayer::init", "LoadingLayer_init", wrapper::LoadingLayer_init, geode::base::get() + 0x319600, Default, LoadingLayer_initHook)
+    CREATE_HOOK_WITH_CHECK_FOR(this->mainClass, "LoadingLayer::loadAssets", "LoadingLayer_loadAssets", wrapper::LoadingLayer_loadAssets, geode::base::get() + 0x319ea0, Default, LoadingLayer_loadAssetsHook)
+    CREATE_HOOK_WITH_CHECK_FOR(this->mainClass, "LoadingLayer::updateProgress", "LoadingLayer_updateProgress", wrapper::LoadingLayer_updateProgress, geode::base::get() + 0x319e20, Default, LoadingLayer_updateProgressHook)
 }